Rework templates so template contents are never parsed
Trying to escape user input strings so the text can be parsed witout any
issues is always going to be a difficult task which will usually have
issues which can be abused. The better solution is to never parse
template strings at all and instead insert the strings directly into the
tree as raw text, never parsing it.

This does change behavior in some ways and requires quoting around
templates in tags, but the end result is more secure and more
predictable.