Add decorations(Map) to component/style builders
This commit introduces a new decorations method to the component and
style builderthat takes in a map of text decorations and their state.
This methodalready exists on the Component interface but was
previously missing on the respective builders.

As the style builder stores the decorations in individual fields and
hence cannot natively work with the map layout, the method was
implemented as a default method on the interface and simply delegates
the map entriess to the normal decorate methods.

The main motivation for the addition of this method is the easy
application of a preset map of decorations.