fix(minimessage): Ignore quotes with no matching closing quote
Don't enter string parsing mode during the first pass when encountering
a quote, but no matching quote exists later on in the message.

Note that this doesn't catch every case of invalid tags causing other
valid tags to be ignored, in this case there could simply be an
unrelated quote elsewhere in the message and the bug would still occur.

The parser could:

  * Only allow quotes after specific characters, such as the colon `:`,
    which is currently the only place where quotes are really valid.
    However, even then, the buggy payload could just include a colon
    before the quote

  * Only allow quotes when inside of a valid tag, as in the test.
    (`<3` isn't a valid introducer for a tag). However, even then, the
    buggy payload could use a character valid for tag. Additionally,
    checking the validity of a tag isn't done by the first parse stage

  * Backtrack when a quote-delimited section doesn't end up being
    inside of a valid tag. That would (presumably) require some pretty
    complicated checks in the first phase, but checking for it in the
    second phase is probably too late, because it'd have to reparse
    something it has already passed by