fix(minimessage): Backtrack when the content of a tag never ends
A tag is opened by `<` and closed by `>`. Upon encountering an opening
angle bracket `<`, the parser enters the "TAG" state, during which it'll
parse quote-delimited sequences. However, if by the end of the input,
a closing angle bracket `>` is never encountered, the parser simply
adds everything as a single text node, including the contents of the
quote-delimited sequences.

These quote delimited sequences could have contained other valid tags
that are now directly added to the output without being parsed as tags.

With this change, when the parser finds itself at the end of the input
in the TAG state, it will backtrack back to the beginning of the `<`
that caused a state change, and parse again from there on, but this time
*not* in the tag state. This means that quotes are effectively ignored
and their contents parsed as normal text

This *still* doesn't account for all corner case, for example:

`<red>foo <tag:"</red>"> bar`. This is an example of a *correct* use of
quotes, and the expected behaviour is for the whole output to be red,
and the actual behaviour matches that, however:

`<red>foo <3:"</red>"> bar`. This is nearly identical to the above, but
<3> isn't a valid tag, which means the red tag should be closed in the
middle of the message. Instead, the message ends up wholly red.

`<red>foo <tag"</red>"> bar`. The same applies here, but instead of an
invalid tag, it is invalid syntax, since quotes should follow a colon.