minimessage: Fix exception with single quote as tag part
The parser tries to un-quote tag parts surrounded by quotes, it does so
by checking if the first character of the tag part is a quote and the
last character is as well. However, things break if that first and last
character are actually the same character! A
StringIndexOutOfBoundsException is thrown later on.

This patch makes it so tag parts consisting of a single quote aren't
tried to be un-quoted, they stay as is

Fixes https://github.com/KyoriPowered/adventure/issues/1011