feat(nbt): TagStringIO Enhancements
Since adventure release 4.21.0, users can now write any BinaryTag to a String; that left a gap reading them.
There is also another case where it may be useful to capture the rest of the string, for example, in a command argument. (Otherwise users would have to group count)
So this commit:
 * Adds support to read any BinaryTag (disallows trailing like the Compound version)
 * Use a writer for any BinaryTag
 * Adds support for reading embedded any BinaryTag, and CompoundBinaryTag (Allows trailing data, through passed Appendable)
 * Adds support to CharBuffer to take the rest of the buffer.
The special cases exist for compound methods so you can immediately throw on the first failed parsing token, even know BinaryTag's can represent a Compound